Telegram API is unstable, catch all errors from calling the API
the telegram API is really unstable from time to time and causes all
sort of errors and not all of them are known. So lets reduce the
`rescue` blocks to only contain the telegram calls, but catch all errors
and only log them.

This still allows errors from outside of telegram to happen and maybe
crash/restart the bot (for example reconnecting to mumble)